Moved the unique code present in the CanDriver.c to main.c such that the CanDriver.c can be moved in the general/util folder. Further Documentation (commenting what each function does) for the CanDriver.c has been added.
